drv:iio:adc:sun4i-gpadc-iio: add A80 thermal sensor
This patch adds support for the A80 ths sensor.

The A80 has 4 sensors and supports interrupts. The interrupt is configured to update
the sensor values every second. The A80 shares some registers with the
integrated GPADC. ACQ0 must be set in the GPADC register with the offset
0x00. In fact the GPADC and the THS use the same register base and also
the same clocks and resets.

Signed-off-by: Philipp Rossak <embed3d@gmail.com>