HACK: clk: rockchip: rk3399: dedicate vpll for vopb and hdmi use
Rockchip PLLs are kown provide the least jitter for
vco rates between 800 MHz and 2 GHz. I added the
rates for VPLL which are used for VOPs dclk and there-
fore HDMI phy in that manner and used the rates which
require the lowest frac divs.
Additionally I added some rates which are useful to
provide additional VESA and non-VESA rates for HDMI
output.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>
Signed-off-by: Alex Bee <knaerzche@gmail.com>