megous/usb-dwc3-core-do-not-use-3.0-clock-when-operating-in(:1)
Original-Subject: usb: dwc3: core: do not use 3.0 clock when operating in 2.0 mode
In the 3.0 device core, if the core is programmed to operate in
2.0 only, then setting the GUCTL1.DEV_FORCE_20_CLK_FOR_30_CLK makes
the internal 2.0(utmi/ulpi) clock to be routed as the 3.0 (pipe)
clock. Enabling this feature allows the pipe3 clock to be not-running
when forcibly operating in 2.0 device mode.

Signed-off-by: Bin Yang <yangbin@rock-chips.com>
Change-Id: I217a380815c21903c1090bd003c1d8ba2fadbe7c
Signed-off-by: Peter Geis <pgwipeout@gmail.com>
X-Armbian: Patch-File: megous/usb-dwc3-core-do-not-use-3.0-clock-when-operating-in
X-Armbian: Patch-File-Counter: 1
X-Armbian: Patch-Rel-Directory: patch/kernel/archive/sunxi-5.15
X-Armbian: Patch-Type: kernel
X-Armbian: Patch-Root-Type: core
X-Armbian: Patch-Sub-Type: common
X-Armbian: Original-Subject: usb: dwc3: core: do not use 3.0 clock when operating in 2.0 mode