add possibility of disabling rk808-rtc
To disable rk808-rtc driver from loading for specific board
add the following stanza to rk808 node in device tree:

	rtc {
		compatible = "rk808-rtc";
		status = "disabled";
	}

This is needed for roc-rk3399-pc plus (a.k.a. Station P1).
Without the change rk808's rtc is initialised and used for time keeping
although there is another rtc (hym8563) that should be actually used.

Signed-off-by: Piotr Szczepanik <piter75@gmail.com>