hwrng: rockchip: Add support for Rockchip RK3588 HW RNG
Note by ColorfulRhino:
The driver 'drivers/char/hw_random/rockchip-rng.c' is already
implemented in mainline, but only for RK3568 as of kernel v6.12.
The hardware RNG in the RK3566 outputs very low entropy, see:
https://lore.kernel.org/linux-rockchip/97dfca058858d7a5d933ddf7a84dba61@manjaro.org/

TODO: Check reliability and quality of entropy for this patched
HW RNG driver on RK3588. If it's bad like in the RK3566 case,
drop this patch.