media: cec: Add MediaTek CECv2 Controller Driver
Add the MediaTek CEC v2 Controller driver, found in SoCs equipped
with the HDMIv2 Controller IP, like MT8195 and MT8188.

In the MT8195 SoC the (single) CEC Controller is shared between
the HDMI-TX and the HDMI-RX controllers, while in MT8188 this is
used only for HDMI-TX (as there is no receiver in the latter).

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>