arm64: dts: mediatek: mt8365-evk : Add overlays for ethernet or hdmi support
On the Mediatek Genio 350-EVK board, Ethernet and HDMI (DPI0) are
sharing pins, and only one functionality can be selected at a time with
a DIP switch. In the board devicetree, this leads to have dpi0 and
ethernet nodes disabled by default.

Add overlay files for the mt8365-evk.dts DT so an user can easily
select a configuration where the Ethernet or HDMI output are enabled
without modifying or rebuilding device tree.

Signed-off-by: Louis-Alexis Eyraud <louisalexis.eyraud@collabora.com>