arm64: dts: rockchip: rk356x: use gpu 200mhz opp as opp-suspend
CLK_GPU is the main clock for the GPU on RK356x, it's typical source
pll can be one of mpll, gpll, cpll or npll. For higher clock rates it
is also possible to use the gpu pvtpll as pll source.

The logic to switch between a normal pll and the pvtpll depending on
rate is handled in TF-A firmware, and exposed to Linux as a scmi clock.
In the device tree this is modeled in the gpu node by using the scmi gpu
clk as the core gpu clk and the normal CLK_GPU as a bus clk.

TF-A will typically change to use normal pll for rates up to 400 MHz and
use pvtpll for 600 MHz or more.

Add opp-suspend to the 200 MHz opp to ensure a normal pll is used when
the gpu device is in PM runtime suspended state.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>