arm64: dts: mediatek: mt8365-evk: Fix touchscreen reset line polarity
Commit 7363096a5a08 ("Input: goodix - remove setting of RST pin to input")
modified the Goodix driver so it no longer tries to set reset pin as
input.

It led the driver to fail to probe during boot time on the Mediatek
Genio 350-EVK board because the reset pin polarity is incorrectly
declared in the board devicetree:
```
[    6.253819] Goodix-TS 1-005d: Error reading 1 bytes from 0x8140: -6
[    6.277884] Goodix-TS 1-005d: Error reading 1 bytes from 0x8140: -6
[    6.305503] Goodix-TS 1-005d: I2C communication failure: -6
```

so fix in the mt8365-evk.dts the reset-gpios property and pinctrl
configuration of the touchscreen node to use the proper polarity value.

Fixes: 7363096a5a08 ("Input: goodix - remove setting of RST pin to input")
Fixes: b28c4af8e44b ("arm64: dts: mt8365-evk: Add goodix touchscreen support")
Signed-off-by: Louis-Alexis Eyraud <louisalexis.eyraud@collabora.com>