arm64: dts: rockchip: Enable the NPU on some rk3588 boards
This series enables the NPU on a few Rockchip rk3588 boards.
In every case here, the required regulator was already in place.
As suggested by Jimmy Hon [1], since the NPU power domain supply is
being described, also remove the always-on property from the regulator.

In Jimmy's words: the commit message for the introduction of the 
pd_npu label mentions how the regulator no longer needs to be always on. 
[2]

I might send a separate series dropping always-on from the same regulator
on the boards that have already added the NPU bits with no changes to 
the regulator. I wonder if I should add Fixes for those.

[1] https://lore.kernel.org/all/CALWfF7KHyKmDcEL9=mYzfCRWc_cymNscoFdeXr8P2frJeww5vA@mail.gmail.com/
[2] https://lore.kernel.org/linux-rockchip/20250721-6-10-rocket-v9-7-77ebd484941e@tomeuvizoso.net/

To: Rob Herring <robh@kernel.org>
To: Krzysztof Kozlowski <krzk+dt@kernel.org>
To: Conor Dooley <conor+dt@kernel.org>
To: Heiko Stuebner <heiko@sntech.de>
Cc: devicetree@vger.kernel.org
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-rockchip@lists.infradead.org
Cc: linux-kernel@vger.kernel.org
Cc: Jimmy Hon <honyuenkwun@gmail.com>
Cc: Tomeu Vizoso <tomeu@tomeuvizoso.net>
Signed-off-by: Ricardo Pardini <ricardo@pardini.net>

---
Changes in v3:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v2: https://lore.kernel.org/r/20260101-arm64-dts-rockchip-rk3588-npu-enablements-v2-0-013cf5d5c39d@pardini.net

Changes in v2:
- Drop the always-on from the vdd_npu_s0 regulator.
- Link to v1: https://lore.kernel.org/r/20251230-arm64-dts-rockchip-rk3588-npu-enablements-v1-0-d38b182a59e3@pardini.net

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 3,
    "change-id": "20251230-arm64-dts-rockchip-rk3588-npu-enablements-1e2830d2e7b0",
    "prefixes": [],
    "history": {
      "v1": [
        "20251230-arm64-dts-rockchip-rk3588-npu-enablements-v1-0-d38b182a59e3@pardini.net"
      ],
      "v2": [
        "20260101-arm64-dts-rockchip-rk3588-npu-enablements-v2-0-013cf5d5c39d@pardini.net"
      ]
    }
  }
}